# Use pubsub bindings in broker ingress