## Proposed Changes

- Directly use pubsub client and bindings in ingress handler. Removes unnecessary overhead and allows for better control over the metrics and trace spans that get emitted. Results in a ~10% performance improvement for small events. For very large events the reduction in overhead is negligible.

## Benchmark Results
```
name                            old time/op  new time/op  delta
IngressHandler/0_bytes-16       11.0µs ± 3%   9.7µs ± 2%  -11.63%  (p=0.000 n=9+10)
IngressHandler/1000_bytes-16    11.2µs ± 1%  10.1µs ± 2%   -9.54%  (p=0.000 n=10+10)
IngressHandler/100000_bytes-16   122µs ± 5%   123µs ± 4%     ~     (p=0.798 n=8+8)
```